--- a/tree_test.go
+++ b/tree_test.go
@@ -34,7 +34,7 @@ func testPath(t *testing.T, tree *node, path string, expectPath string, expected
 	expectCatchAll := strings.Contains(expectPath, "/*")
 
 	t.Log("Testing", path)
-	n, paramList := tree.search(path[1:])
+	n, paramList := tree.search(path[1:], nil)
 	if expectPath != "" && n == nil {
 		t.Errorf("No match for %s, expected %s", path, expectPath)
 		return
@@ -294,13 +294,49 @@ func TestPanics(t *testing.T) {
 	twoPathPanic(":abc/ggg", ":def/ggg")
 }
 
+type TestMatcher struct {
+	match bool
+}
+
+func (fm *TestMatcher) isMatch(value interface{}) bool {
+	return fm.match
+}
+
+func TestFalseMatcher(t *testing.T) {
+	tree := &Tree{}
+	err := tree.Add("/some/path", 1)
+	if err != nil {
+		t.Error(err)
+	}
+
+	v, _ := tree.Lookup("/some/path", &TestMatcher{false})
+
+	if v != nil {
+		t.Error("failed, no match expected for false matcher")
+	}
+}
+
+func TestTrueMatcher(t *testing.T) {
+	tree := &Tree{}
+	err := tree.Add("/some/path", 1)
+	if err != nil {
+		t.Error(err)
+	}
+
+	v, _ := tree.Lookup("/some/path", &TestMatcher{true})
+
+	if v == nil {
+		t.Error("failed, match expected for true matcher")
+	}
+}
+
 func BenchmarkTreeNullRequest(b *testing.B) {
 	b.ReportAllocs()
 	tree := &node{path: "/"}
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		tree.search("")
+		tree.search("", nil)
 	}
 }
 
@@ -311,7 +347,7 @@ func BenchmarkTreeOneStatic(b *testing.B) {
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		tree.search("abc")
+		tree.search("abc", nil)
 	}
 }
 
@@ -322,6 +358,6 @@ func BenchmarkTreeOneParam(b *testing.B) {
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		tree.search("abc")
+		tree.search("abc", nil)
 	}
 }