--- a/tree_test.go
+++ b/tree_test.go
@@ -298,7 +298,7 @@ type TestMatcher struct {
 	match bool
 }
 
-func (fm *TestMatcher) isMatch(value interface{}) bool {
+func (fm *TestMatcher) IsMatch(value interface{}) bool {
 	return fm.match
 }