--- a/tree_test.go
+++ b/tree_test.go
@@ -34,7 +34,7 @@ func testPath(t *testing.T, tree *node, path string, expectPath string, expected
 	expectCatchAll := strings.Contains(expectPath, "/*")
 
 	t.Log("Testing", path)
-	n, paramList := tree.search(path[1:], tm)
+	n, paramList, _ := tree.search(path[1:], tm)
 	if expectPath != "" && n == nil {
 		t.Errorf("No match for %s, expected %s", path, expectPath)
 		return
@@ -298,8 +298,8 @@ type TestMatcher struct {
 	match bool
 }
 
-func (fm *TestMatcher) Match(value interface{}) bool {
-	return fm.match
+func (fm *TestMatcher) Match(value interface{}) (bool, interface{})	 {
+	return fm.match, value
 }
 
 func TestFalseMatcher(t *testing.T) {
@@ -309,7 +309,7 @@ func TestFalseMatcher(t *testing.T) {
 		t.Error(err)
 	}
 
-	v, _ := tree.LookupMatcher("/some/path", &TestMatcher{false})
+	v, _, _ := tree.LookupMatcher("/some/path", &TestMatcher{false})
 
 	if v != nil {
 		t.Error("failed, no match expected for false matcher")
@@ -323,7 +323,7 @@ func TestTrueMatcher(t *testing.T) {
 		t.Error(err)
 	}
 
-	v, _ := tree.LookupMatcher("/some/path", &TestMatcher{true})
+	v, _, _ := tree.LookupMatcher("/some/path", &TestMatcher{true})
 
 	if v == nil {
 		t.Error("failed, match expected for true matcher")