This is implementation of fix for https://github.com/zalando/skipper/issues/116

Lookup just have Matcher and when node founded, checks that it is matched by matcher.

This implementation have BC break and performance degradation.

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkTreeNullRequest-4     5.46          6.32          +15.75%
BenchmarkTreeOneStatic-4       23.9          24.7          +3.35%
BenchmarkTreeOneParam-4        21.4          23.5          +9.81%
```

I think all another implementations will have similar degradation, mostly it is related to add new argument to function signatures.