@aryszka 2016-04-11
thanks for checking the benchmarks. I find the performance degradation acceptable. Especially in Skipper's use case, I expect it to become insignificant. It would be nicer though, if it was possible to avoid the degradation in those cases when the Matcher is not used during the lookup, but I don't see currently a 'nice' solution for that :)