@ewgRa 2016-04-13
Agree with comments. Will commit fixes soon.