@ewgRa 2016-04-13
Done. I've added also init function to initialize trueMatcher

New bench:

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkTreeNullRequest-4     5.46          7.13          +30.59%
BenchmarkTreeOneStatic-4       23.9          25.3          +5.86%
BenchmarkTreeOneParam-4        21.4          23.3          +8.88%
```

Degradation on NullRequest I think acceptable. Bench shows that on real cases there is not so big degradation.