@ewgRa 2016-04-13
Please do not merge this PR without solution here https://github.com/zalando/skipper/pull/133#discussion_r59264934. Since solution can have changes in Matcher interface and better to merge final solution and do not fight later with broken BC.