@aryszka 2016-04-14
as the previous comment mentions, i agree to find a solution for [skipper/#133](https://github.com/zalando/skipper/pull/133#discussion_r59264934) first.