@ewgRa 2016-04-18
Done. 

1. Test passed, but I miss before this https://github.com/zalando/pathmux/pull/2/files#diff-37cf0fd7a164303538a2d22d965646dfL272 at line 272. I think for catchAll also Matcher must be called? And seems there is no tests for this.

2.Latest changes finally have such performance degradation for pathmux:

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkTreeNullRequest-4     5.61          8.79          +56.68%
BenchmarkTreeOneStatic-4       23.9          28.5          +19.25%
BenchmarkTreeOneParam-4        21.4          24.6          +14.95%
```

I just think about this return Match (bool, interface) last commit change. pathmux tests looks for me a little bit synthetic. I have an idea, that if revert it and return in skipper this matchLeaves double call - it can work even faster, than with such value return from Matcher. But can't proof it since skipper benchmarks require a lot of memory to run.