@aryszka 2016-04-19
i ran the tests in skipper, here is a comparison:

```
[master/1]
1 route               1000000          1000 ns/op
100 routes            1000000          1749 ns/op
10000 routes          1000000          1904 ns/op
1000000 routes         500000          3549 ns/op

[master/2]
1 route               2000000           918 ns/op
100 routes            1000000          1268 ns/op
10000 routes          1000000          1830 ns/op
1000000 routes         500000          3533 ns/op

[master/3]
1 route               2000000           952 ns/op
100 routes            1000000          1217 ns/op
10000 routes          1000000          1915 ns/op
1000000 routes         300000          3469 ns/op

--------------------------------------------------

[fix/1]
1 route               1000000          1100 ns/op
100 routes            1000000          1404 ns/op
10000 routes           500000          2005 ns/op
1000000 routes         500000          3914 ns/op

[fix/2]
1 route               1000000          1121 ns/op
100 routes            1000000          1399 ns/op
10000 routes          1000000          1997 ns/op
1000000 routes         500000          3908 ns/op

[fix/3]
1 route               1000000          1120 ns/op
100 routes            1000000          1416 ns/op
10000 routes           500000          2108 ns/op
1000000 routes         500000          3923 ns/op
```

discussed it with @mo-gr, and we both were on the opinion that it is ok from Skipper's standpoint (which is the greater good for us :) )

regarding the tests and benchmarks in pathmux, note that this is fork httptreemux, and we weren't changing the logic sofar. Probably, the benchmarks wouldn't show so big degradation here, if they were run against a larger tree where the majority of the recursion steps wouldn't call the Matcher.Match function. Or, it is also possible to create a separate recursion for the simple Lookup case, too. But I think that can be done in a separate PR.

However, now this LGTM, and if the doc is updated, then i'd like to merge this.