@ewgRa 2016-04-19
done, doc blocks added